+++ js/web-components/keydown-action-service.js
@@ -5,7 +5,7 @@
  * and allows the [enter] key of the element to trigger a [click()] event on the element
  * specified in the selector.
  *
- * The term "service" is used here because all elments on the page with the
+ * The term "service" is used here because all elements on the page with the
  * attribute will have an event assigned to them so rather than rendering content
  * this Web Component provides a "service" for other elements on the page.
  *
@@ -19,23 +19,15 @@
  *     <input type="text" data-bind="name" data-enter-key-click-selector="button">
  */
 
-/* Validates with both [jshint] and [eslint] */
-/* jshint esversion:8 */
-/* eslint-env browser, es6 */
-
-import { usingWebComponentsPolyfill } from './utils.js';
+import { WebComponentService } from './WebComponentService.js';
 
 /**
- * Shadow DOM for Custom Elements
+ * Click Event that get's attached to [data-enter-key-click-selector] elements
+ *
+ * @param {KeyboardEvent} e
  */
-const shadowTmpl = document.createElement('template');
-shadowTmpl.innerHTML = `
-    <style> :host { display:none; } </style>
-    <slot></slot>
-`;
-
 function enterToClick(e) {
-    if (e.keyCode !== 13) {
+    if (e.key !== 'Enter') {
         return;
     }
     const selector = e.target.getAttribute('data-enter-key-click-selector');
@@ -51,36 +43,13 @@ function enterToClick(e) {
 }
 
 /**
- * Define class for the <keydown-action-service> element
- */
-class KeydownActionService extends HTMLElement {
-    constructor() {
-        super();
-        if (usingWebComponentsPolyfill()) {
-            return;
-        }
-        const shadowRoot = this.attachShadow({mode: 'open'});
-        shadowRoot.appendChild(shadowTmpl.content.cloneNode(true));
-    }
-
-    connectedCallback() {
-        this.setEvents(true);
-    }
-
-    disconnectedCallback() {
-        this.setEvents(false);
-    }
-
-    setEvents(add) {
-        const elements = document.querySelectorAll('[data-enter-key-click-selector]');
-        const event = (add ? 'addEventListener' : 'removeEventListener');
-        for (const el of elements) {
-            el[event]('keydown', enterToClick);
-        }
-    }
-}
-
-/**
  * Add <keydown-action-service> element to the page
  */
-window.customElements.define('keydown-action-service', KeydownActionService);
+window.customElements.define('keydown-action-service', class KeydownActionService extends WebComponentService {
+    refresh(rootElement) {
+        const elements = rootElement.querySelectorAll('[data-enter-key-click-selector]');
+        for (const el of elements) {
+            el.addEventListener('keydown', enterToClick);
+        }
+    }
+});